+++ sample/.env.example
@@ -1,5 +1,3 @@
 AUTH0_DOMAIN={DOMAIN}
 AUTH0_CLIENT_ID={CLIENT_ID}
 AUTH0_CLIENT_SECRET={CLIENT_SECRET}
-AUTH0_CALLBACK_URL=http://localhost:3000/auth/auth0/callback
-AUTH0_AUDIENCE=
