+++ sample/.ruby-version
@@ -1,1 +1,1 @@
-2.7.1
+2.7
