+++ sample/app/controllers/auth0_controller.rb
@@ -1,10 +1,10 @@
 class Auth0Controller < ApplicationController
   def callback
-    # OmniAuth places the User Profile information (retrieved by omniauth-auth0) in request.env['omniauth.auth'].
-    # In this tutorial, you will store that info in the session, under 'userinfo'.
+    # OmniAuth stores the informatin returned from Auth0 and the IdP in request.env['omniauth.auth'].
+    # In this sample, you will pull the raw_info supplied from the id_token
     # If the id_token is needed, you can get it from session[:userinfo]['credentials']['id_token'].
-    # Refer to https://github.com/auth0/omniauth-auth0#auth-hash for complete information on 'omniauth.auth' contents.
-    session[:userinfo] = request.env['omniauth.auth']
+    # Refer to https://github.com/auth0/omniauth-auth0#authentication-hash for complete information on 'omniauth.auth' contents.
+    session[:userinfo] = request.env['omniauth.auth']['extra']['raw_info']
 
     redirect_to '/dashboard'
   end
@@ -13,5 +13,26 @@ class Auth0Controller < ApplicationController
   # passing the error message in the 'message' request param.
   def failure
     @error_msg = request.params['message']
+  end
+
+  def logout
+    reset_session
+    redirect_to logout_url
+  end
+
+  private
+  AUTH0_CONFIG = Rails.application.config_for(:auth0)
+
+  def logout_url
+    request_params = {
+      returnTo: root_url,
+      client_id: AUTH0_CONFIG['auth0_client_id']
+    }
+
+    URI::HTTPS.build(host: AUTH0_CONFIG['auth0_domain'], path: '/v2/logout', query: to_query(request_params)).to_s
+  end
+
+  def to_query(hash)
+    hash.map { |k, v| "#{k}=#{CGI.escape(v)}" unless v.nil? }.reject(&:nil?).join('&')
   end
 end
