--- sample/app/controllers/logout_controller.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-class LogoutController < ApplicationController
-  include LogoutHelper
-  def logout
-    reset_session
-    redirect_to logout_url.to_s
-  end
-end
