+++ sample/app/helpers/auth0_helper.rb
@@ -1,0 +1,2 @@
+module Auth0Helper
+end
