--- sample/app/helpers/logout_helper.rb
@@ -1,19 +1,0 @@
-# frozen_string_literal: true
-module LogoutHelper
-  def logout_url
-    domain = Rails.application.config.auth0['auth0_domain']
-    client_id = Rails.application.config.auth0['auth0_client_id']
-    request_params = {
-        returnTo: root_url,
-        client_id: client_id
-    }
-
-    URI::HTTPS.build(host: domain, path: '/v2/logout', query: to_query(request_params))
-  end
-
-  private
-
-  def to_query(hash)
-    hash.map { |k, v| "#{k}=#{URI.escape(v)}" unless v.nil? }.reject(&:nil?).join('&')
-  end
-end
