+++ sample/config/routes.rb
@@ -1,16 +1,11 @@
 Rails.application.routes.draw do
-  get 'dashboard/show'
-
-  get '/' => 'home#show'
+  root 'home#show'
 
   get '/dashboard' => 'dashboard#show'
-  get '/logout' => 'logout#logout'
-
   get '/auth/auth0/callback' => 'auth0#callback'
   get '/auth/failure' => 'auth0#failure'
+  get '/auth/logout' => 'auth0#logout'
   get '/auth/redirect' => 'auth0#redirect'
-
-  root 'home#show'
 
   # The priority is based upon order of creation: first created -> highest priority.
   # See how all your routes lay out with "rake routes".
